requirements.txt
notion-client
python-dotenv
requests
pandas